requirements.txt
allure-pytest==2.8.36
allure-python-commons==2.8.36
apipkg==1.5
appnope==0.1.2
attrs==20.3.0
backcall==0.2.0
browserstack-local
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
execnet==1.8.0
flaky==3.7.0
idna==2.10
iniconfig==1.1.1
ipdb==0.13.6
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
packaging==20.8
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.16
ptyprocess==0.7.0
py==1.10.0
Pygments==2.8.1
pyparsing==2.4.7
pytest==6.2.1
pytest-forked==1.3.0
pytest-order==0.10.0
pytest-xdist==2.2.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
